requirements.txt
Flask
Flask-SQLAlchemy
Flask-JWT-Extended
Flask-Bcrypt
Flask-CORS
flasgger
psycopg2-binary
requests
oauthlib
pyOpenSSL
python-dotenv
regex
Flask-Mail
openai
pillow
flask-migrate